HediffHUD: Added hediffs now list available operations
Hediffs will now show a list of all operations available that install a prosthetic or implant.
Operation list will be added regardless of if a hediff exists, but will also hide operations that add any already-existing part, as well as parts that can't be installed due to a missing parent part (e.g. hand talon with bionic arm).